# adapter/adapter.rest/adapter.rest.v2/src/main/webapp/api-docs/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "https://github.com/learning-layers/SocialSemanticServer/",
    "version" : "v2",
    "title" : "Social Semantic Server REST API",
    "termsOfService" : "https://github.com/learning-layers/",
    "contact" : {
      "email" : "dtheiler#tugraz$at"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host" : "localhost:8080",
  "basePath" : "/sss.adapter.rest.v2/rest",
...
  "paths" : {
...
    "/livingdocs/livingdocs/filtered/{livingDoc}" : {
      "post" : {
        "tags" : [ "livingdocs" ],
        "summary" : "retrieve living doc for a user",
        "description" : "",
        "operationId" : "livingDocFilteredGet",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "livingDoc",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/SSLivingDocGetRet"
            }
          }
        }
      }
    },
    "/livingdocs/livingdocs/{livingDoc}" : {
      "get" : {
        "tags" : [ "livingdocs" ],
        "summary" : "retrieve a living doc",
        "description" : "",
        "operationId" : "livingDocGet",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "livingDoc",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/SSLivingDocGetRet"
            }
          }
        }
      },
      "put" : {
        "tags" : [ "livingdocs" ],
        "summary" : "update a living doc",
        "description" : "",
        "operationId" : "livingDocUpdate",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "livingDoc",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/SSLivingDocUpdateRet"
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "livingdocs" ],
        "summary" : "remove a living doc",
        "description" : "",
        "operationId" : "livingDocRemove",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "livingDoc",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/SSLivingDocRemoveRet"
            }
          }
        }
      }
    },
...
  },
  "definitions" : {
...
    "SSEntity" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "string"
        },
        "label" : {
          "type" : "string"
        },
        "creationTime" : {
          "type" : "integer",
          "format" : "int64"
        },
        "type" : {
          "type" : "string"
        },
        "author" : {
          "$ref" : "#/definitions/SSEntity"
        },
        "description" : {
          "type" : "string"
        },
        "circleTypes" : {
          "type" : "array",
          "items" : {
            "type" : "string"
          }
        },
        "entries" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/SSEntity"
          }
        },
        "attachedEntities" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/SSEntity"
          }
        },
        "comments" : {
          "type" : "array",
          "items" : {
            "type" : "string"
          }
        },
        "overallRating" : {
          "$ref" : "#/definitions/SSEntityA"
        },
        "tags" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/SSEntity"
          }
        },
        "categories" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/SSEntity"
          }
        },
        "discs" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/SSEntity"
          }
        },
        "userEvents" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/SSEntity"
          }
        },
        "thumb" : {
          "$ref" : "#/definitions/SSEntity"
        },
        "file" : {
          "$ref" : "#/definitions/SSEntity"
        },
        "flags" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/SSEntityA"
          }
        },
        "users" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/SSEntity"
          }
        },
        "entities" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/SSEntity"
          }
        },
        "circles" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/SSEntity"
          }
        },
        "locations" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/SSEntity"
          }
        },
        "read" : {
          "type" : "boolean",
          "default" : false
        },
        "likes" : {
          "$ref" : "#/definitions/SSEntityA"
        },
        "profilePicture" : {
          "$ref" : "#/definitions/SSEntity"
        },
        "activityPage" : {
          "$ref" : "#/definitions/SSQueryResultPage"
        },
        "messagesPage" : {
          "$ref" : "#/definitions/SSQueryResultPage"
        },
        "collsPage" : {
          "$ref" : "#/definitions/SSQueryResultPage"
        }
      }
    },
    "SSEntityA" : {
      "type" : "object"
    },
...
    "SSLivingDocGetRet" : {
      "type" : "object",
      "required" : [ "op" ],
      "properties" : {
        "op" : {
          "type" : "string"
        },
        "livingDoc" : {
          "$ref" : "#/definitions/SSEntity"
        }
      }
    },
    "SSLivingDocRemoveRet" : {
      "type" : "object",
      "required" : [ "op" ],
      "properties" : {
        "op" : {
          "type" : "string"
        },
        "livingDoc" : {
          "type" : "string"
        }
      }
    },
    "SSLivingDocUpdateRet" : {
      "type" : "object",
      "required" : [ "op" ],
      "properties" : {
        "op" : {
          "type" : "string"
        },
        "livingDoc" : {
          "type" : "string"
        }
      }
    },
...
    "SSQueryResultPage" : {
      "type" : "object",
      "properties" : {
        "pagesID" : {
          "type" : "string"
        },
        "pageNumber" : {
          "type" : "integer",
          "format" : "int32"
        },
        "pagesCount" : {
          "type" : "integer",
          "format" : "int32"
        },
        "entities" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/SSEntity"
          }
        }
      }
    },
...
  }
}